Replace sbyte and byte with ints where applicable

+++ src/NFugue/Integration/MusicXml/Internals/MidiInstrument.cs
@@ -6,10 +6,10 @@
         public string Channel { get; }
         public string Name { get; }
         public string Bank { get; }
-        public byte Program { get; }
+        public int Program { get; }
         public string Unpitched { get; set; }
 
-        public MidiInstrument(string id, string channel, string name, string bank, byte program, string unpitched)
+        public MidiInstrument(string id, string channel, string name, string bank, int program, string unpitched)
         {
             Id = id;
             Channel = channel;
