+++ src/NFugue/Staccato/Functions/PitchWheelFunction.cs
@@ -15,8 +15,8 @@ namespace NFugue.Staccato.Functions
             var splittedParams = parameters.Split(',');
             if (splittedParams.Length == 2)
             {
-                context.Parser.OnPitchWheelParsed(sbyte.Parse(splittedParams[0].Trim()),
-                    sbyte.Parse(splittedParams[1].Trim()));
+                context.Parser.OnPitchWheelParsed(int.Parse(splittedParams[0].Trim()),
+                    int.Parse(splittedParams[1].Trim()));
             }
             else if (splittedParams.Length == 1)
             {
