+++ src/NFugue/Staccato/Functions/PolyPressureFunction.cs
@@ -14,8 +14,8 @@ namespace NFugue.Staccato.Functions
             var splittedParams = parameters.Split(',');
             if (splittedParams.Length == 2)
             {
-                context.Parser.OnPolyphonicPressureParsed(sbyte.Parse(splittedParams[0].Trim()),
-                    sbyte.Parse(splittedParams[1].Trim()));
+                context.Parser.OnPolyphonicPressureParsed(int.Parse(splittedParams[0].Trim()),
+                    int.Parse(splittedParams[1].Trim()));
             }
         }
     }
